bare-metal-blinky: New application to blink an LED at 1 Hz
An exercise in peripheral configuration. The 32.678 low-frequency RC oscillator
(LFRCO) provides a clock source to the low-energy timer peripheral (LETIMER).

As the LETIMER0 underflows, an internal peripheral reflex system (PRS)  signal
is toggled. The PRS signal is then asynchronously delivered to the GPIO pin that
drives LED0.

While it may have been possible to route the LETIMER's underflow output directly
to a pin using the "GPIO alternate modes", the PRS provides much greater
flexability in routing. Further, no processing time is devoted toggling the LED,
and the timer will continue to run down to the deepest processor sleep (EM4).