Add Adafruit Neopixel Controller application / driver
A Neopixel control signal is generated using the LDMA, USART, and a lookup table
to control the entire string of LEDs with minimal processing time.

For comparison, an arduino microcontroller that bit-bangs a GPIO within a critical
section would need ~2 ms critical sections to change light level.

With a DMA, the CPU can hand off data transfer duties to the low-power peripherals
while it goes to make an espresso.

Next steps:
  1. Bring in FreeRTOS
  2. Extend to 10-bits with dithering
  3. Basic BLE GATT Server applications to be controlled from a phone
  4. Fade logic